finish setup of pages with CA
