+++ src/app/search/[id]/booking/profesional/date/page.tsx
@@ -1,15 +1,10 @@
-import React from "react";
-import { CalendarSelect } from "./components/calendar-select";
+import React, { Suspense } from "react";
+import { SelectDatePage } from "@/layers/search/page/compnay-id/page/date/page";
 
-export default function SelectDatePage() {
+export default function page() {
   return (
-    <div className="flex  w-full h-full">
-      <section className="space-y-2 w-full h-full flex flex-col">
-        <p className="text-gray-500 font-semibold">Seleccionar una fecha</p>
-        <div className="flex-grow">
-          <CalendarSelect />
-        </div>
-      </section>
-    </div>
+    <Suspense>
+      <SelectDatePage />
+    </Suspense>
   );
 }
