+++ src/app/search/[id]/booking/profesional/page.tsx
@@ -1,59 +1,10 @@
-"use client";
-
-import React, { useEffect, useState } from "react";
-import SelectMemberCard from "./components/select-member-card";
-import { usePathname, useSearchParams, redirect } from "next/navigation";
-import { getServicesMembers } from "@/lib/clienta-actions";
-import { IMember } from "@/interfaces/member.iterface";
-import { LoaderSpinner } from "@/components/common/loader-spinner";
+import React, { Suspense } from "react";
+import { ProfesionalPage } from "@/layers/search/page/compnay-id/page/profesional/page";
 
 export default function Page() {
-  const [users, setUsers] = useState<IMember[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const searchParams = useSearchParams();
-  const pathname = usePathname();
-  const serviceId = searchParams.get("service");
-
-  useEffect(() => {
-    const fetchMembers = async () => {
-      if (serviceId) {
-        try {
-          setIsLoading(true);
-          const members = await getServicesMembers(serviceId);
-          setUsers(members);
-        } catch (error) {
-          console.error("Error fetching service members:", error);
-        } finally {
-          setIsLoading(false);
-        }
-      }
-    };
-
-    fetchMembers();
-  }, [serviceId]);
-
-  if (!serviceId) {
-    const redirectionURL = pathname
-      .split("/")
-      .splice(0, pathname.split("/").length - 1)
-      .join("/");
-    return redirect(redirectionURL);
-  }
-
   return (
-    <div className="space-y-4 ">
-      {isLoading ? (
-        <p className="flex items-center gap-2">
-          Cargando
-          <LoaderSpinner />
-        </p>
-      ) : users.length === 0 ? (
-        "No hay profesionales disponibles"
-      ) : (
-        users.map((member) => (
-          <SelectMemberCard member={member} key={member.id} />
-        ))
-      )}
-    </div>
+    <Suspense>
+      <ProfesionalPage />
+    </Suspense>
   );
 }
