+++ src/layers/customers/page/index.ts
@@ -1,1 +1,2 @@
 export * from "./customer-page";
+export * from "./customer-id-page";
