+++ src/layers/search/page/compnay-id/components/appointment-card-info.tsx
@@ -1,7 +1,7 @@
 import { ICompany } from "@/interfaces";
 import React from "react";
 
-export default function AppointmentInfo({ company }: { company: ICompany }) {
+export function AppointmentInfo({ company }: { company: ICompany }) {
   return (
     <aside className="w-1/3 p-4 bg-white rounded-lg shadow-md h-[80vh]">
       <h2 className="text-xl font-bold mb-4">{company.name}</h2>
