+++ src/layers/search/page/compnay-id/page/date/page/index.ts
@@ -1,0 +1,1 @@
+export * from "./date-page";
