+++ src/layers/search/page/compnay-id/page/index.ts
@@ -1,0 +1,2 @@
+export * from "./search-company-id";
+export * from "./booking-page";
