+++ src/layers/search/page/compnay-id/page/profesional/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./select-member-card";
