+++ src/layers/search/page/compnay-id/page/profesional/page/index.ts
@@ -1,0 +1,1 @@
+export * from "./profesional-page";
